# docs/openapi.yaml
openapi: 3.0.3
info:
  title: gosso - SSO Authentication Server
  description: |
    A production-ready Single Sign-On (SSO) authentication server built with Go.

    Supports:
    - Username/password login with MFA (TOTP + backup codes)
    - OAuth 2.0 Authorization Server (authorization code, refresh token, client credentials)
    - OpenID Connect Provider (ID token, discovery, JWKS, userinfo)
    - Social login (Google, GitHub, WeChat)
    - Admin account management
    - Email/phone verification and password reset
  version: 1.0.0
  contact:
    name: rushairer
    url: https://github.com/rushairer/gosso
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT
servers:
  - url: http://localhost:8080
    description: Local development server
...
paths:
...
  /api/auth/refresh:
    post:
      tags: [Auth]
      summary: Refresh access token
      description: Exchanges a valid refresh token for a new access token and refresh token pair.
      operationId: refreshToken
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RefreshTokenRequest"
      responses:
        "200":
          description: Token refreshed
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response_TokenPair"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
  /api/auth/passkey/register/begin:
    post:
      tags: [Passkey]
      summary: Begin passkey registration
      description: Start WebAuthn credential creation ceremony. Returns PublicKeyCredentialCreationOptions.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: Registration options
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/passkey/register/complete:
    post:
      tags: [Passkey]
      summary: Complete passkey registration
      description: Complete WebAuthn credential creation ceremony with the browser's attestation response.
      security:
        - BearerAuth: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
      responses:
        "200":
          description: Passkey registered
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/passkey/login/begin:
    post:
      tags: [Passkey]
      summary: Begin passkey login
      description: Start WebAuthn authentication ceremony. If account_id is provided, uses regular login; otherwise uses discoverable login.
      requestBody:
        content:
          application/json:
            schema:
              type: object
              properties:
                account_id:
                  type: string
                  description: Optional account ID for non-discoverable login
      responses:
        "200":
          description: Login options with request_id
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
  /api/auth/passkey/login/complete:
    post:
      tags: [Passkey]
      summary: Complete passkey login
      description: Complete WebAuthn authentication ceremony with the browser's assertion response.
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required: [request_id]
              properties:
                request_id:
                  type: string
      responses:
        "200":
          description: Login successful
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/passkeys:
    get:
      tags: [Passkey]
      summary: List passkeys
      description: List all registered passkeys for the authenticated account.
      security:
        - BearerAuth: []
      responses:
        "200":
          description: List of passkeys
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "401":
          $ref: "#/components/responses/Unauthorized"
  /api/auth/passkeys/{id}:
    delete:
      tags: [Passkey]
      summary: Delete passkey
      description: Delete a registered passkey (ownership check enforced).
      security:
        - BearerAuth: []
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: string
      responses:
        "200":
          description: Passkey deleted
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Response"
        "400":
          $ref: "#/components/responses/BadRequest"
        "401":
          $ref: "#/components/responses/Unauthorized"
...
components:
...
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
      description: JWT access token obtained from login, token, or social login endpoints
    BasicAuth:
      type: http
      scheme: basic
      description: HTTP Basic Auth with client_id as username and client_secret as password
    CSRFToken:
      type: apiKey
      in: header
      name: X-CSRF-Token
      description: |
        CSRF protection via double-submit cookie pattern.
        1. GET any endpoint to receive `csrf_token` cookie
        2. Read cookie value and send as `X-CSRF-Token` header on POST/PUT/DELETE
        3. Not required when using `Authorization: Bearer` header (JWT auth)

  # ── Parameters ──
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 400
            message: invalid request body
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/ErrorResponse"
          example:
            code: 401
            message: unauthorized
...
  schemas:
...
    Response:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          description: Response data (type varies by endpoint)
...
    ErrorResponse:
      type: object
      properties:
        code:
          type: integer
          example: 400
        message:
          type: string
          example: error message
...
    RefreshTokenRequest:
      type: object
      required: [refresh_token]
      properties:
        refresh_token:
          type: string
...
    Response_TokenPair:
      type: object
      properties:
        code:
          type: integer
          example: 200
        message:
          type: string
          example: success
        data:
          type: object
          properties:
            access_token:
              type: string
            refresh_token:
              type: string
            token_type:
              type: string
              example: Bearer
            expires_in:
              type: integer
              example: 900
...